Changed the scriptcs entrypoint to handle the case where a script is given with no file extension a little differently, as discussed in https://github.com/scriptcs/scriptcs/issues/515.
The change basically mean that we'll watch out for the case when there is no script specified, and that we only extend the script name if its given and don't have an extension, and that we'll make the addition transparent in terms of what code will handle things like a missing file (it's now deferred to the command). This assures we don't ever risk messing with other commands.