Unit tests for #629 and a few updates
Changed how I initialized the ILineProcessor list.
Tests to verify the LineProcessor<T>() function adds the type to the
Overrides dictionary when called from ModuleConfiguration.
Also found and fixed a problem where ServiceOverrides._this was null
when using the non default constructor.  Unit test for this is in
ModuleConfigurationTests.TheLineProcessorMethod.ShouldReturnTheModuleConfiguration
Added RunTimeServicestests for verifying that the default line
processors and custom processors get registered when the build container
method is called.